+++ spin2/CHANGELOG.md
@@ -18,6 +18,17 @@ Possible next additions:
 - Add new-file templates as Snippets
 - Add additional Snippets as the community identifies them
 
+## [2.2.11] 2023-12-??
+
+Update P2 Only
+
+- Add recognition of "auto" on debug scope display declarations
+- Add Semantic highlight color change for byte(), word(), and long() method overrides
+- Add recognition of byte(), word(), and long() method names to provide method vs. storage type hover text
+- Add recognition of {Spin2_v4XX} format Spin Language Requirement directive
+- Emit any languge directive in use to generated interface documentation
+- Add support for lstring() when {Spin2_v43} is specified
+
 ## [2.2.10] 2023-12-24
 
 Update P1 and P2